+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changes in version 0.9-7.1 (2016-xx-xx)
 
+* OPTICS: added a predecessor correction step that is used by 
+    the ELKI implementation.  
+* Added hullplot to plot a scatter plot with added convex cluster hulls.
+
 # Changes in version 0.9-7 (2016-04-14)
 
 * OPTICSXi is now implemented (thanks to Matt Piekenbrock).